Move solution writing step into global build handler
