+++ ProjectSettings/ProjectSettings.asset
@@ -270,7 +270,7 @@ PlayerSettings:
   - m_BuildTarget: 
     m_Icons:
     - serializedVersion: 2
-      m_Icon: {fileID: 0}
+      m_Icon: {fileID: 2800000, guid: 74ebfa89af37aba43b08982d7f97eb7d, type: 3}
       m_Width: 128
       m_Height: 128
       m_Kind: 0
